add subfolder page for blog

+++ createhtml.sh
@@ -4,8 +4,23 @@ pre='pre.html'
 post='post.html'
 for page in src/Pages/* 
 do 
-  filename=$dest${page##*/}
-  cat $pre > $filename
-  cat $page >> $filename
-  cat $post >> $filename
+  if [ -f $page ]
+  then
+    filename=$dest${page##*/}
+    cat $pre > $filename
+    cat $page >> $filename
+    cat $post >> $filename
+  fi
+done
+mkdir -p predist/Blog
+dest='predist/Blog/'
+for page in src/Pages/Blog/* 
+do 
+  if [ -f $page ]
+  then
+    filename=$dest${page##*/}
+    cat $pre > $filename
+    cat $page >> $filename
+    cat $post >> $filename
+  fi
 done
